Refactor: Extract diagnostics and rejection tracking to src/analysis/diagnostics.rs (#271)
Extract diagnostic tracking and rejection reason structures from
implementation.rs (~900 lines) to a dedicated diagnostics.rs module.

## Structures Extracted

### Core Adapter
- `RecordCacheProvider` struct - Adapter for focus impact calculation
- `compute_impact_scores_for_discounting()` function

### Synapse Rejection Tracking
- `RejectionReason` enum - Why a synapse candidate was rejected
- `RejectionDetail` struct - Detailed rejection information
- `ThresholdContext` struct - Threshold calculation context
- `TargetDiagnosticEntry` struct - Per-target diagnostic
- `TargetDiagnostics` struct - Collection of target diagnostics

### Neuron Rejection Tracking
- `NeuronRejectionDetail` struct - Why neuron candidate rejected
- `NeuronDiagnosticEntry` struct - Per-neuron diagnostic
- `NeuronDiagnostics` struct - Collection of neuron diagnostics

### Target Data Structures
- `TargetData` struct - Target neuron data for sample building
- `TargetMap` struct - Pre-built target map for efficient sample building

### Focus Target Filtering
- `FocusTargetFilterResult` struct - Result of filtering focus targets
- `filter_focus_targets_for_neuron_analysis()` function

### Validation
- `require_unique_focus()` function - Validate focus targets are unique

## Benefits

Diagnostics provide visibility into why candidates were rejected:
- `NEAT_AI_DISCOVERY_VERBOSE=1` enables detailed logging
- JSON responses include `diagnostics` arrays
- Critical for debugging "no candidates found" situations

This continues the refactoring effort from Issue #185 to break down the
implementation.rs monolith into focused submodules.
